Moved some code around for consistent style.
- Deprecated import of JsonImplicits in favor of importing play.api.libs.json.ops._
