+++ src/main/scala/play/api/libs/json/scalacheck/Counted.scala
