+++ src/main/scala/play/api/libs/json/scalacheck/DurationGenerators.scala
@@ -1,10 +1,9 @@
 package play.api.libs.json.scalacheck
 
-import org.scalacheck.{Gen, Arbitrary}
+import org.scalacheck.{Arbitrary, Gen}
 
 import scala.concurrent.duration._
 import scala.concurrent.duration.ops._
-
 import scala.language.implicitConversions
 
 trait DurationGenerators {
