+++ src/main/scala/play/api/libs/json/scalacheck/Width.scala
